worker/requirements.txt
celery==4.4.7
psycopg2-binary==2.9.5
redis==4.3.6
moviepy==1.0.3
psycopg2-binary==2.9.5

